requirements.txt
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
cffi==1.14.5
class-registry==2.1.2
colorama==0.4.4
cryptography==3.4.6
Django==3.1.7
django-cors-headers==3.7.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-formtools==2.2
django-heroku==0.3.1
djangorestframework==3.12.2
django-rest-knox==4.1.0
dj-database-url==0.5.0
filters==1.3.2
filters-django==1.0.5
gunicorn==20.0.4
importlib-metadata==3.7.0
iniconfig==1.1.1
more-itertools==8.7.0
mysqlclient==2.0.3
packaging==20.9
Pillow==8.1.1
pip==21.0.1
pluggy==0.13.1
psycopg2==2.8.6
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
python-http-client==3.3.2
pytz==2021.1
regex==2020.11.13
reportlab==3.5.59
sendgrid==3.5.0
sendgrid-django==4.2.0
setuptools==54.0.0
six==1.15.0
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
toml==0.10.2
wcwidth==0.2.5
whitenoise==5.2.0
zipp==3.4.0
typing-extensions==3.7.4.3
